+++ podcast/podcast_service/src/main/java/ismaeldivita/audioma/podcast/service/itunes/model/ItunesPodcast.kt
@@ -1,12 +1,12 @@
 package ismaeldivita.audioma.podcast.service.itunes.model
 
 data class ItunesPodcast(
-    val id: Int,
+    val id: Long,
     val title: String,
     val artistName: String,
     val rssUrl: String,
     val artworkList: List<ItunesArtwork>,
-    val primaryGenreId: Int,
-    val genreListId: List<Int>,
+    val primaryGenreId: Long,
+    val genreListId: List<Long>,
     val explicit: Boolean
 )
